# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion RH API — ARTF",
        "description": "API REST de gestion RH (intégration administrative des agents).\n\n## Authentification\n\n1. Appeler `POST /api/login` avec email / mot de passe\n2. Copier le `token` retourné\n3. Cliquer **Authorize** en haut à droite\n4. Saisir : `Bearer {token}` (ou uniquement le token selon le client)\n\nComptes de test (seeders) :\n- `admin@arft.cg` / `Admin@2026`\n- `rh@arft.cg` / `Rh@2026`\n\n## Convention de réponse\n\n```json\n{ \"data\": {}, \"message\": \"...\" }\n```\n\nErreurs : `{ \"success\": false, \"message\": \"...\", \"errors\": {} }`",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API (variable L5_SWAGGER_CONST_HOST)"
        }
    ],
    "paths": {
        "/api/integration/dossiers/{dossier}/actes": {
            "get": {
                "tags": [
                    "Intégration — Actes"
                ],
                "summary": "Lister les actes d'un dossier",
                "operationId": "listActesDossier",
                "parameters": [
                    {
                        "name": "dossier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ActeAdministratif"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Intégration — Actes"
                ],
                "summary": "Générer un acte administratif (manuel)",
                "operationId": "genererActeAdministratif",
                "parameters": [
                    {
                        "name": "dossier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type_acte"
                                ],
                                "properties": {
                                    "type_acte": {
                                        "type": "string",
                                        "example": "decision_recrutement"
                                    },
                                    "contenu": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ActeAdministratifResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/integration/dossiers": {
            "get": {
                "tags": [
                    "Intégration — Dossiers"
                ],
                "summary": "Liste des dossiers d'intégration",
                "operationId": "listDossiersIntegration",
                "parameters": [
                    {
                        "name": "statut",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "BROUILLON"
                    },
                    {
                        "name": "type_integration_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DossierIntegrationListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Intégration — Dossiers"
                ],
                "summary": "Créer un dossier d'intégration",
                "operationId": "storeDossierIntegration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DossierIntegrationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DossierIntegrationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error422"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/integration/dossiers/{id}/soumettre": {
            "post": {
                "tags": [
                    "Intégration — Dossiers"
                ],
                "summary": "Soumettre le dossier pour étude RH",
                "operationId": "soumettreDossierIntegration",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Soumis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DossierIntegrationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error404"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/integration/dossiers/{id}/passer-en-etude-rh": {
            "post": {
                "tags": [
                    "Intégration — Dossiers"
                ],
                "summary": "Prendre en charge le dossier (RH)",
                "operationId": "passerEnEtudeRHDossier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "En étude RH",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DossierIntegrationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error401": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Non authentifié. Veuillez vous connecter."
                    }
                },
                "type": "object"
            },
...
            "ActeAdministratifResponse": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/ActeAdministratif"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "DossierIntegrationListResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DossierIntegration"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Sanctum obtenu via POST /api/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}